🐛 BUGFIX: set placeholder values for Person
