+++ types/uclapi.d.ts
@@ -53,6 +53,14 @@ export interface TimetableEvent {
   start_time?: string,
 }
 
+
+export interface Person {
+  name?: string,
+  status?: string,
+  department?: string,
+  email?: string,
+}
+
 export type NestedNavigator<T> = {
   screen: keyof T,
   params?: T[keyof T],
